fetch/metadata/generated/css-font-face.sub.tentative.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/css-font-face.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>HTTP headers on request for CSS font-face</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  const params = {
  };

  let count = 0;

  function induceRequest(t, url) {
    const id = `el-${count += 1}`;
    const style = document.createElement('style');
    style.appendChild(document.createTextNode(`
      @font-face {
        font-family: wpt-font-family${id};
        src: url(${url});
      }
      #el-${id} {
        font-family: wpt-font-family${id};
      }
    `));
    const div = document.createElement('div');
    div.setAttribute('id', 'el-' + id);
    div.appendChild(style);
    div.appendChild(document.createTextNode('x'));
    document.body.appendChild(div);

    t.add_cleanup(() => div.remove());

    return document.fonts.ready;
  }

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['sameOrigin', 'httpOrigin'], params))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
      });
  }, 'sec-fetch-site - HTTPS downgrade (header not sent)');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['httpOrigin', 'sameOrigin'], params))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
      });
  }, 'sec-fetch-site - HTTPS upgrade');

  promise_test((t) => {
    const key = '{{uuid()}}';

    return induceRequest(t, makeRequestURL(key, ['sameOrigin', 'httpOrigin', 'sameOrigin'], params))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
      });
  }, 'sec-fetch-site - HTTPS downgrade-upgrade');
  </script>
  </body>
</html>
